requirements.txt
addict==2.4.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
cachetools==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
dj-database-url==2.2.0
Django==4.2.4
django-bulma==0.8.3
django-crispy-forms==2.3
django-environ==0.8.1
django-storages==1.14.4
django-widget-tweaks==1.4.12
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
fonttools==4.42.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
nida==0.1.6
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==10.0.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydyf==0.10.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.16.0
pyphen==0.14.0
pypng==0.20220715.0
python-barcode==0.15.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.6
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.16
XlsxWriter==3.2.0
zopfli==0.2.3
django-ninja
